chore: rename NavHeader file to index for a better understand the structure
